update so that ceph video storage is not deleted after use, error checking with ceph bucket

+++ flask-video-serving-service/update.py
@@ -31,7 +31,7 @@ s3r = boto3.resource('s3',
 
 
 
-def build_video(): 
+def build_video(currSeg): 
     
     my_bucket = s3r.Bucket(name=s3_bucket)
 
@@ -39,17 +39,28 @@ def build_video():
       os.mkdir("segments")
     ffmpeg_log = open('segments/ffmpeg_log.txt', 'w')
     print(my_bucket.objects.all())
+    
     for my_bucket_object in my_bucket.objects.all():
-        print("Getting Segment from Ceph: ",my_bucket_object.key)
         
-        s3.download_file(s3_bucket,my_bucket_object.key,"segments/"+my_bucket_object.key)
-        os.system("ffmpeg -i segments/"+my_bucket_object.key+" -s 640x360 -hls_flags delete_segments+append_list+omit_endlist -hls_list_size 30 -f hls segments/out.m3u8")
-        s3.delete_object(Bucket=s3_bucket,Key=my_bucket_object.key)
-        os.remove("segments/"+my_bucket_object.key)
+        print("Getting Segment from Ceph: ",my_bucket_object.key)
+        filename = my_bucket_object.key
+        num = filename.split("-")
+        currNum = int(num[1].split(".")[0])
+        
+        if currNum > currSeg:
+            s3.download_file(s3_bucket,my_bucket_object.key,"segments/"+my_bucket_object.key)
+            os.system("ffmpeg -i segments/"+my_bucket_object.key+" -s 640x360 -hls_flags delete_segments+append_list+omit_endlist -hls_list_size 30 -f hls segments/out.m3u8")
+            #s3.delete_object(Bucket=s3_bucket,Key=my_bucket_object.key)
+            os.remove("segments/"+my_bucket_object.key)
+            currSeg = currNum
+        else: continue
+
+    return currSeg
 
 
 if __name__ == '__main__':
 
+    currSeg = -1
     while(True):
-        build_video()
+        currSeg = build_video(currSeg)
         time.sleep(5)
